# [OptimizationMOI] improve performance of Jacobian and Hessian accesses