@odow 2024-01-16
Format-check seems unrelated. Also failing on `master`: https://github.com/SciML/Optimization.jl/actions/runs/7493571366/job/20399569099