@odow 2024-01-16
This makes little difference to the `Optimization.AutoSparseReverseDiff()` case because the main cost is computing the Hessian sparsity

<img width="809" alt="image" src="https://github.com/SciML/Optimization.jl/assets/8177701/dc01d897-4c6c-4952-ad38-7e737d77bd5b">