@odow 2024-01-16
It's just part of MOI, sooo it's already a package?

The main blocker is that you'd need to convert the symbolic form into MOI's expression format. It also operates on a single model with all the constraints, not function-by-function.